requirements.txt
pandas
numpy
requests
tqdm
pillow
joblib
python-dotenv
opencv-python
face_recognition
pytesseract
ultralytics
networkx
scikit-learn
torch
torchvision
web3
ipfshttpclient
flickrapi==2.4.0
py-solc-x